requirements.txt
streamlit
tensorflow
pandas
numpy
scikit-learn
pillow
opencv-python-headless
pymongo
requests
matplotlib
plotly
python-dotenv
bcrypt
google-generativeai
openai>=1.0.0
apscheduler>=3.10.4
